added twfe simulation
This simulation uses two way fixed effects to estimate the causal effect of a treatment that varies by time period across groups when the causal effect is constant and when there are dynamics.  It shows the bias that results when using twoway fixed effects.  It then implements Callaway and Sant’anna’s 2019 estimator for estimating group/time average treatment effects.  It is missing, though, an aggregation and the bootstrapping for inference currently. It is named baker.do after Andrew Baker who wrote a similar simulation in R.